searchsmith/requirements.txt
# --- Core LangChain Framework ---
# Essential packages for the LangChain library
langchain
langchain-core
langchain-community

# --- Vector Database ---
# For creating and managing the local vector store
chromadb

# For using local and API-based models from Hugging Face
langchain-huggingface

# --- Model & ML Libraries ---
# The core library for running local sentence embedding models
sentence-transformers

# The main machine learning framework, required by sentence-transformers
torch

# A core library from Hugging Face, a dependency for many models like Qwen
transformers

# A dependency for specific models like Qwen
einops

# --- Utilities ---
# For loading environment variables (like API keys) from a .env file
python-dotenv

# For creating progress bars, useful for long-running scripts
tqdm

# For handling Markdown formatted text if needed
markdown


fastapi
uvicorn
